backend/warafi/requirements.txt
asgiref>=3.3.1,<4
mysqlclient==2.1.0
PyPDF2
langchain
openai
djangorestframework-simplejwt
Django>=3.0,<6.0
certifi==2020.12.5
django-elasticsearch-dsl
elasticsearch==7.10.1
elasticsearch-dsl>=7.0.0,<8.0.0
python-dateutil==2.8.1
pytz==2020.5
sqlparse==0.4.1
urllib3
kor
six
django-cors-headers
django-environ